Move classify tag to contrib
