+++ meinberlin/apps/budgeting/templates/meinberlin_budgeting/includes/proposal_list_item.html
@@ -1,4 +1,4 @@
-{% load i18n item_tags moderatorfeedback_tags humanize contrib_tags %}
+{% load i18n item_tags humanize contrib_tags %}
 
 <li class="list-item list-item--squashed">
     <div class="list-item__stats">
