+++ meinberlin/apps/budgeting/templates/meinberlin_budgeting/proposal_detail.html
@@ -1,5 +1,5 @@
 {% extends "meinberlin_mapideas/mapidea_detail.html" %}
-{% load rules item_tags moderatorfeedback_tags i18n humanize contrib_tags %}
+{% load rules item_tags i18n humanize contrib_tags %}
 
 {% block additional_labels %}
     {{ block.super }}
