+++ meinberlin/apps/contrib/templatetags/contrib_tags.py
@@ -1,7 +1,11 @@
+import re
+import unicodedata
+
 from django import template
 from django.forms.utils import flatatt
 from django.template import defaultfilters
 from django.template.loader import render_to_string
+from django.utils.encoding import force_text
 from django.utils.safestring import mark_safe
 
 register = template.Library()
@@ -66,3 +70,22 @@ def html_date(value, displayfmt=None, datetimefmt='c', **kwargs):
                                                    datetime,
                                                    displaydate)
     return mark_safe(result)
+
+
+@register.filter
+def classify(value):
+    """
+    Create a valid CSS class name from a value.
+
+    Converts to ASCII. Converts spaces to dashes. Removes characters that
+    aren't alphanumerics, underscores, or hyphens.
+    Also strips leading and trailing whitespace.
+    """
+    if value is None:
+        return 'NONE'
+
+    value = force_text(value)
+    value = unicodedata.normalize('NFKD', value) \
+        .encode('ascii', 'ignore').decode('ascii')
+    value = re.sub('[^\w\s-]', '', value).strip()
+    return mark_safe(re.sub('[-\s]+', '-', value))
